Allow symlinks to be used in vold.fstab
With device-tree kernels the sysfs path for SD cards varies
depending on device-tree layout and the actual card used. It is
therefore impossible to specify the sysfs path for the device
in vold.fstab.

However, there exists a symlink to the device which is created under a
generic path like '/sys/block/mmcblk0/device'. To use such a path with
vold we need to convert the symlinked path to the real path, which this
patch does.

Change-Id: Ifc810593a8fa390f875762eecd5c3413e62c1d16
Signed-off-by: Jon Medhurst <tixy@linaro.org>